founder-ai/requirements.txt
# LLM
google-genai>=1.0.0

# Web Search
ddgs>=9.0.0

# Web UI
fastapi>=0.115.0
uvicorn>=0.32.0

# Env
python-dotenv>=1.0.0
